[MOV] website_sale: move s_products_recently_viewed to dedicated file
Follows: #42067

Part of: https://github.com/odoo/odoo/pull/45334
task-2118851
