+++ addons/website_sale/__manifest__.py
@@ -20,8 +20,9 @@
         'views/sale_order_views.xml',
         'views/crm_team_views.xml',
         'views/templates.xml',
-        'views/snippets.xml',
+        'views/snippets/snippets.xml',
         'views/snippets/s_products_searchbar.xml',
+        'views/snippets/s_products_recently_viewed.xml',
         'views/res_config_settings_views.xml',
         'views/digest_views.xml',
         'views/website_sale_visitor_views.xml',
