+++ addons/website_sale/static/src/snippets/s_products_recently_viewed/000.scss
@@ -1,0 +1,45 @@
+.s_wsale_products_recently_viewed {
+    .o_carousel_product_card {
+        .o_carousel_product_card_img_top {
+            object-fit: scale-down;
+            @include media-breakpoint-down(sm) {
+                height: 12rem;
+            }
+            @include media-breakpoint-up(md) {
+                height: 8rem;
+            }
+            @include media-breakpoint-up(lg) {
+                height: 12rem;
+            }
+        }
+        .o_carousel_product_img_link:hover + .o_carousel_product_remove {
+            display: block;
+        }
+    }
+
+    .o_carousel_product_card_wrap {
+        @include media-breakpoint-up(sm) {
+            float: left;
+        }
+    }
+
+    .o_carousel_product_control {
+        top: percentage(1/3);
+        bottom: percentage(1/3);
+        width: 2rem;
+        border-radius: 5px;
+        background-color: $o-enterprise-primary-color;
+    }
+
+    .o_carousel_product_remove {
+        position: absolute;
+        display: none;
+        cursor: pointer;
+        right: 5%;
+        top: 5%;
+    }
+
+    .o_carousel_product_remove:hover {
+        display: block;
+    }
+}
