+++ addons/website_sale/views/snippets/snippets.xml
@@ -1,21 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-<template id="s_products_recently_viewed" name="Products Recently Viewed">
-    <section class="s_wsale_products_recently_viewed d-none pt24 pb24" style="min-height: 400px;">
-        <div class="container">
-            <div class="alert alert-info alert-dismissible rounded-0 fade show d-print-none css_non_editable_mode_hidden o_not_editable">
-                This is a preview of the recently viewed products by the user.<br/>
-                Once the user has seen at least one product this snippet will be visible.
-                <button type="button" class="close" data-dismiss="alert" aria-label="Close"> × </button>
-            </div>
-
-            <h3 class="text-center mb32">Recently viewed Products</h3>
-            <div class="slider o_not_editable"/>
-        </div>
-    </section>
-</template>
-
 <template id="snippets" inherit_id="website.snippets" name="e-commerce snippets">
     <xpath expr="//t[@id='sale_recently_viewed_product_hook']" position="replace">
         <t t-snippet="website_sale.s_products_recently_viewed" t-thumbnail="/website_sale/static/src/img/snippets_thumbs/s_products_recently_viewed.png"/>
